LS: Move all go to definition logic into the `ide::navigation` module
commit-id:07f289ac
