+++ crates/cairo-lang-language-server/src/ide/mod.rs
@@ -1,4 +1,5 @@
 pub mod completion;
 pub mod formatter;
 pub mod hover;
+pub mod navigation;
 pub mod semantic_highlighting;
