+++ crates/cairo-lang-language-server/src/ide/navigation/mod.rs
@@ -1,0 +1,1 @@
+pub mod goto_definition;
